+++ kaspr/resources/__init__.py
@@ -1,6 +1,6 @@
+from .kasprtable import KasprTable
 from .kaspragent import KasprAgent
 from .kasprwebview import KasprWebView
 from .kasprapp import KasprApp
-from .kasprscheduler import KasprScheduler
 
-__all__ = ["KasprApp", "KasprScheduler", "KasprAgent", "KasprWebView"]
+__all__ = ["KasprApp", "KasprAgent", "KasprWebView", "KasprTable"]