+++ kaspr/resources/kasprapp.py
@@ -53,7 +53,7 @@ from kubernetes.client import (
 )
 
 from kaspr.resources.base import BaseResource
-from kaspr.resources import KasprAgent, KasprWebView
+from kaspr.resources import KasprAgent, KasprWebView, KasprTable
 from kaspr.common.models.labels import Labels
 
 
@@ -61,8 +61,11 @@ class KasprApp(BaseResource):
     """Kaspr App kubernetes resource."""
 
     KIND = "KasprApp"
+    GROUP_NAME = "kaspr.io"
     GROUP_VERSION = "v1alpha1"
     COMPONENT_TYPE = "app"
+    PLURAL_NAME = "kasprapps"
+    KASPR_APP_NAME_LABEL = "kaspr.io/app"
     WEB_PORT_NAME = "http"
     KASPR_CONTAINER_NAME = "kaspr"
 
@@ -114,12 +117,15 @@ class KasprApp(BaseResource):
     _pod_template: V1PodTemplateSpec = None
     _agent_pod_volumes: List[V1Volume] = None
     _webview_pod_volumes: List[V1Volume] = None
+    _table_pod_volumes: List[V1Volume] = None
 
     # Reference to agent resources
     agents: List[KasprAgent] = None
     webviews: List[KasprWebView] = None
+    tables: List[KasprTable] = None
     _webviews_hash: str = None
     _agents_hash: str = None
+    _tables_hash: str = None
 
     # TODO: Templates allow customizing k8s behavior
     template_service_account: ResourceTemplate
@@ -138,20 +144,23 @@ class KasprApp(BaseResource):
         kind: str,
         namespace: str,
         component_type: str,
+        labels: Optional[Dict[str, str]] = None,
     ):
         component_name = KasprAppResources.component_name(name)
-        labels = Labels.generate_default_labels(
+        _labels = Labels.generate_default_labels(
             name,
             kind,
             component_name,
             component_type,
             self.KASPR_OPERATOR_NAME,
         )
+        _labels.update(labels or {})
+        _labels.update({self.KASPR_APP_NAME_LABEL: name})
         super().__init__(
             cluster=name,
             namespace=namespace,
             component_name=component_name,
-            labels=labels,
+            labels=_labels,
         )
 
     @classmethod
@@ -199,6 +208,11 @@ class KasprApp(BaseResource):
         """Apply webview resources to the app."""
         self.webviews = webviews
         self._webviews_hash = None  # reset hash
+
+    def with_tables(self, tables: List[KasprTable]):
+        """Apply table resources to the app."""
+        self.tables = tables
+        self._tables_hash = None # reset hash
 
     def fetch(self, name: str, namespace: str):
         """Fetch actual KasprApp in kubernetes."""
@@ -347,6 +361,10 @@ class KasprApp(BaseResource):
         if self.webviews:
             env_vars.append(V1EnvVar(name="WEBVIEWS_HASH", value=self._webviews_hash))
 
+        # include tables hash
+        if self.tables:
+            env_vars.append(V1EnvVar(name="TABLES_HASH", value=self._tables_hash))
+
         return env_vars
 
     def prepare_kafka_credentials_env_dict(self) -> Dict[str, str]:
@@ -380,8 +398,7 @@ class KasprApp(BaseResource):
             env_for("web_port"): str(self.web_port),
             env_for("data_dir"): self.data_dir_path,
             env_for("table_dir"): self.table_dir_path,
-            env_for("definitions_dir"): self.definitions_dir_path,
-            env_for("topic_prefix"): self.topic_prefix,
+            env_for("definitions_dir"): self.definitions_dir_path
         }
         _envs = {**config_envs}
         _envs.update(overrides)
@@ -412,12 +429,28 @@ class KasprApp(BaseResource):
                 )
             )
         return volume_mounts
+    
+    def prepare_table_volume_mounts(self) -> List[V1VolumeMount]:
+        volume_mounts = []
+        for table in self.tables if self.tables else []:
+            volume_mounts.append(
+                V1VolumeMount(
+                    name=table.volume_mount_name,
+                    mount_path=self.prepare_table_mount_path(table),
+                    read_only=True,
+                    sub_path=table.file_name,
+                )
+            )
+        return volume_mounts
 
     def prepare_agent_mount_path(self, agent: KasprAgent) -> str:
         return f"{self.definitions_dir_path}/{agent.file_name}"
 
     def prepare_webview_mount_path(self, webview: KasprWebView) -> str:
         return f"{self.definitions_dir_path}/{webview.file_name}"
+    
+    def prepare_table_mount_path(self, table: KasprTable) -> str:
+        return f"{self.definitions_dir_path}/{table.file_name}"
 
     def prepare_volume_mounts(self) -> List[V1VolumeMount]:
         volume_mounts = []
@@ -429,6 +462,7 @@ class KasprApp(BaseResource):
                 ),
                 *self.prepare_agent_volume_mounts(),
                 *self.prepare_webview_volume_mounts(),
+                *self.prepare_table_volume_mounts()
             ]
         )
         return volume_mounts
@@ -464,6 +498,7 @@ class KasprApp(BaseResource):
         volumes = []
         volumes.extend(self.prepare_agent_volumes())
         volumes.extend(self.prepare_webview_volumes())
+        volumes.extend(self.prepare_table_volumes())
         return volumes
 
     def prepare_agent_volumes(self) -> List[V1Volume]:
@@ -490,6 +525,22 @@ class KasprApp(BaseResource):
                         name=webview.config_map_name,
                         items=[
                             V1KeyToPath(key=webview.file_name, path=webview.file_name)
+                        ],
+                    ),
+                )
+            )
+        return volumes
+    
+    def prepare_table_volumes(self) -> List[V1Volume]:
+        volumes = []
+        for table in self.tables if self.tables else []:
+            volumes.append(
+                V1Volume(
+                    name=table.volume_mount_name,
+                    config_map=V1ConfigMapVolumeSource(
+                        name=table.config_map_name,
+                        items=[
+                            V1KeyToPath(key=table.file_name, path=table.file_name)
                         ],
                     ),
                 )
@@ -722,6 +773,20 @@ class KasprApp(BaseResource):
         ]
         kopf.adopt(children)
 
+    def search(self, namespace: str, apps: List[str] = None):
+        """Search for KasprApps in kubernetes."""
+        label_selector = (
+            ",".join(f"{self.KASPR_APP_NAME_LABEL}={app}" for app in apps) if apps else None
+        )
+        return self.list_custom_objects(
+            self.custom_objects_api,
+            namespace=namespace,
+            group=self.GROUP_NAME,
+            version=self.GROUP_VERSION,
+            plural=self.PLURAL_NAME,
+            label_selector=label_selector,
+        )
+    
     def agents_status(self) -> Dict:
         """Return status of all agents."""
         return [agent.info() for agent in self.agents]
@@ -729,6 +794,10 @@ class KasprApp(BaseResource):
     def webviews_status(self) -> Dict:
         """Return status of all webviews."""
         return [webview.info() for webview in self.webviews]
+    
+    def tables_status(self) -> Dict:
+        """Return status of all tables."""
+        return [table.info() for table in self.tables]
 
     @cached_property
     def version(self) -> KasprVersion:
@@ -753,10 +822,6 @@ class KasprApp(BaseResource):
     @cached_property
     def definitions_dir_path(self):
         return getattr(self.config, "definitions_dir", self.DEFAULT_DEFINITIONS_DIR)
-
-    @cached_property
-    def topic_prefix(self):
-        return getattr(self.config, "topic_prefix", f"{self.cluster}.")
 
     @cached_property
     def apps_v1_api(self) -> AppsV1Api:
@@ -879,6 +944,12 @@ class KasprApp(BaseResource):
         return self._webview_pod_volumes
 
     @cached_property
+    def table_pod_volumes(self) -> List[V1Volume]:
+        if self._table_pod_volumes is None:
+            self._table_pod_volumes = self.prepare_table_volumes()
+        return self._table_pod_volumes
+    
+    @cached_property
     def agents_hash(self) -> str:
         """Hash of all agents."""
         if self._agents_hash is None:
@@ -899,3 +970,14 @@ class KasprApp(BaseResource):
                 else None
             )
         return self._webviews_hash
+    
+    @cached_property
+    def tables_hash(self) -> str:
+        """Hash of all tables."""
+        if self._tables_hash is None:
+            self._tables_hash = (
+                self.compute_hash("".join(table.hash for table in self.tables))
+                if self.tables
+                else None
+            )
+        return self._tables_hash