+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.4.2
+--
+* Improvements to how resource `status` is updated
+* Improvements to reconciliation
+
+* 
+## 0.4.0
+--
+* Support KasprTable CRD
+* Use camelCase to define app configuration keys
+* Removed `topic_prefix` app setting
+* Removed `KasprScheduler` CRD
+
 ## 0.3.8
 --
 * Rename KasprApp resources with `-app` suffix